feat(hostd,runtime): fold eBPF vsock egress telemetry into core crates
Remove the standalone mvm-ebpf-egress crate and move the Aya loader,
eBPF program, and observability target metadata into the existing
core crates:

- mvm-runtime: VmObservabilityTarget sidecar persisted in mode.json,
  plus build_from_start_config and read_pid_file helpers.
- mvm-hostd: EbpfTelemetryManager wired into Supervisor::launch/stop,
  with platform stubs for macOS/Windows and a real Aya path on Linux.
- mvm-core: vsock egress counters added to the global metrics snapshot.

The eBPF program itself now lives under crates/mvm-hostd/ebpf/ and is
built separately with nightly + bpf-linker so the main workspace stays
on stable Rust.

Focused unit tests, workspace check, and workspace all-target Clippy
pass on the host. The real eBPF object build and an end-to-end
launch→attach→detach test remain for the Linux builder VM.

Refs: #2211
