+++ crates/mvm-hostd/ebpf/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "mvm-ebpf-egress-ebpf"
+name = "mvm-hostd-ebpf"
 version = "0.18.0"
 edition = "2021"
 
@@ -8,7 +8,7 @@ aya-ebpf = "0.2"
 aya-log-ebpf = "0.2"
 
 [[bin]]
-name = "mvm-ebpf-egress"
+name = "mvm-hostd-ebpf"
 path = "src/main.rs"
 
 [profile.dev]
